fix: refactor find entity func
